OpenSora v1.1 training

+++ examples/opensora_hpcai/opensora/models/layers/rotary_embedding.py
@@ -36,7 +36,7 @@ def apply_rotary_emb(
     ), f"feature dimension {t.shape[-1]} is not of sufficient size to rotate in all the positions {rot_dim}"
 
     t_left, t, t_right = t[..., :start_index], t[..., start_index:end_index], t[..., end_index:]
-    t = (t * freqs.cos() * scale) + (rotate_half(t) * freqs.sin() * scale)
+    t = (t * freqs.cos().astype(t.dtype) * scale) + (rotate_half(t) * freqs.sin().astype(t.dtype) * scale)
     return ops.cat((t_left, t, t_right), axis=-1)
